+++ Assets/Mu3LibraryAssets/Samples~/Sample_Template/Scripts/Runtime/Sample_Addressables/AddressableGroupKeys.cs
@@ -12,6 +12,16 @@
         {
             public const string Name = "Localization-Assets-Shared";
             
+            public static readonly string[] AllNames = new string[]
+            {
+                "TestStringTable Shared Data",
+            };
+            
+            public static readonly string[] AllAddresses = new string[]
+            {
+                "Assets/Mu3LibraryAssets/Samples~/Sample_Localization/Localization/Tables/TestStringTable Shared Data.asset",
+            };
+            
             public static class AssetsMu3LibraryAssetsSamplesSampleLocalizationLocalizationTablesTestStringTableSharedDataAsset
             {
                 public const string Name = "TestStringTable Shared Data";
@@ -20,7 +30,10 @@
                 {
                     public const string BaseDownload = "base-download";
                     
-                    public static readonly string[] All = new string[] { "base-download" };
+                    public static readonly string[] All = new string[]
+                    {
+                        "base-download",
+                    };
                 }
             }
         }
@@ -28,6 +41,20 @@
         public static class LocalizationLocales
         {
             public const string Name = "Localization-Locales";
+            
+            public static readonly string[] AllNames = new string[]
+            {
+                "English (en)",
+                "Japanese (ja)",
+                "Korean (ko)",
+            };
+            
+            public static readonly string[] AllAddresses = new string[]
+            {
+                "English (en)",
+                "Japanese (ja)",
+                "Korean (ko)",
+            };
             
             public static class EnglishEn
             {
@@ -38,7 +65,11 @@
                     public const string BaseDownload = "base-download";
                     public const string Locale = "Locale";
                     
-                    public static readonly string[] All = new string[] { "base-download", "Locale" };
+                    public static readonly string[] All = new string[]
+                    {
+                        "base-download",
+                        "Locale",
+                    };
                 }
             }
 
@@ -51,7 +82,11 @@
                     public const string BaseDownload = "base-download";
                     public const string Locale = "Locale";
                     
-                    public static readonly string[] All = new string[] { "base-download", "Locale" };
+                    public static readonly string[] All = new string[]
+                    {
+                        "base-download",
+                        "Locale",
+                    };
                 }
             }
 
@@ -64,7 +99,11 @@
                     public const string BaseDownload = "base-download";
                     public const string Locale = "Locale";
                     
-                    public static readonly string[] All = new string[] { "base-download", "Locale" };
+                    public static readonly string[] All = new string[]
+                    {
+                        "base-download",
+                        "Locale",
+                    };
                 }
             }
         }
@@ -72,6 +111,16 @@
         public static class LocalizationStringTablesEnglishEn
         {
             public const string Name = "Localization-String-Tables-English (en)";
+            
+            public static readonly string[] AllNames = new string[]
+            {
+                "TestStringTable_en",
+            };
+            
+            public static readonly string[] AllAddresses = new string[]
+            {
+                "TestStringTable_en",
+            };
             
             public static class TestStringTableEn
             {
@@ -82,7 +131,11 @@
                     public const string BaseDownload = "base-download";
                     public const string LocaleEn = "Locale-en";
                     
-                    public static readonly string[] All = new string[] { "base-download", "Locale-en" };
+                    public static readonly string[] All = new string[]
+                    {
+                        "base-download",
+                        "Locale-en",
+                    };
                 }
             }
         }
@@ -90,6 +143,16 @@
         public static class LocalizationStringTablesJapaneseJa
         {
             public const string Name = "Localization-String-Tables-Japanese (ja)";
+            
+            public static readonly string[] AllNames = new string[]
+            {
+                "TestStringTable_ja",
+            };
+            
+            public static readonly string[] AllAddresses = new string[]
+            {
+                "TestStringTable_ja",
+            };
             
             public static class TestStringTableJa
             {
@@ -100,7 +163,11 @@
                     public const string BaseDownload = "base-download";
                     public const string LocaleJa = "Locale-ja";
                     
-                    public static readonly string[] All = new string[] { "base-download", "Locale-ja" };
+                    public static readonly string[] All = new string[]
+                    {
+                        "base-download",
+                        "Locale-ja",
+                    };
                 }
             }
         }
@@ -108,6 +175,16 @@
         public static class LocalizationStringTablesKoreanKo
         {
             public const string Name = "Localization-String-Tables-Korean (ko)";
+            
+            public static readonly string[] AllNames = new string[]
+            {
+                "TestStringTable_ko",
+            };
+            
+            public static readonly string[] AllAddresses = new string[]
+            {
+                "TestStringTable_ko",
+            };
             
             public static class TestStringTableKo
             {
@@ -118,7 +195,11 @@
                     public const string BaseDownload = "base-download";
                     public const string LocaleKo = "Locale-ko";
                     
-                    public static readonly string[] All = new string[] { "base-download", "Locale-ko" };
+                    public static readonly string[] All = new string[]
+                    {
+                        "base-download",
+                        "Locale-ko",
+                    };
                 }
             }
         }
@@ -126,6 +207,18 @@
         public static class TestPack
         {
             public const string Name = "TestPack";
+            
+            public static readonly string[] AllNames = new string[]
+            {
+                "SceneThumbnails",
+                "TestImage03",
+            };
+            
+            public static readonly string[] AllAddresses = new string[]
+            {
+                "Assets/Mu3LibrarySamples/Sample_Template/Images/SceneThumbnails",
+                "TestImage03",
+            };
             
             public static class AssetsMu3LibrarySamplesSampleTemplateImagesSceneThumbnails
             {
@@ -135,11 +228,30 @@
                 {
                     public const string DownloadAll = "download-all";
                     
-                    public static readonly string[] All = new string[] { "download-all" };
+                    public static readonly string[] All = new string[]
+                    {
+                        "download-all",
+                    };
                 }
 
                 public static class Assets
                 {
+                    public static readonly string[] AllNames = new string[]
+                    {
+                        "Thumbnail_AudioManager",
+                        "Thumbnail_AudioManagerFor3DSFX",
+                        "Thumbnail_InputSystemManager",
+                        "Thumbnail_Localization",
+                    };
+                    
+                    public static readonly string[] AllAddresses = new string[]
+                    {
+                        "Assets/Mu3LibrarySamples/Sample_Template/Images/SceneThumbnails/Thumbnail_AudioManager.png",
+                        "Assets/Mu3LibrarySamples/Sample_Template/Images/SceneThumbnails/Thumbnail_AudioManagerFor3DSFX.png",
+                        "Assets/Mu3LibrarySamples/Sample_Template/Images/SceneThumbnails/Thumbnail_InputSystemManager.png",
+                        "Assets/Mu3LibrarySamples/Sample_Template/Images/SceneThumbnails/Thumbnail_Localization.png",
+                    };
+                    
                     public static class ThumbnailAudioManagerPng
                     {
                         public const string Name = "Thumbnail_AudioManager";
@@ -148,7 +260,10 @@
                         {
                             public const string DownloadAll = "download-all";
                             
-                            public static readonly string[] All = new string[] { "download-all" };
+                            public static readonly string[] All = new string[]
+                            {
+                                "download-all",
+                            };
                         }
                     }
 
@@ -160,7 +275,10 @@
                         {
                             public const string DownloadAll = "download-all";
                             
-                            public static readonly string[] All = new string[] { "download-all" };
+                            public static readonly string[] All = new string[]
+                            {
+                                "download-all",
+                            };
                         }
                     }
 
@@ -172,7 +290,10 @@
                         {
                             public const string DownloadAll = "download-all";
                             
-                            public static readonly string[] All = new string[] { "download-all" };
+                            public static readonly string[] All = new string[]
+                            {
+                                "download-all",
+                            };
                         }
                     }
 
@@ -184,7 +305,10 @@
                         {
                             public const string DownloadAll = "download-all";
                             
-                            public static readonly string[] All = new string[] { "download-all" };
+                            public static readonly string[] All = new string[]
+                            {
+                                "download-all",
+                            };
                         }
                     }
                 }
@@ -198,7 +322,10 @@
                 {
                     public const string TestImage = "test-image";
                     
-                    public static readonly string[] All = new string[] { "test-image" };
+                    public static readonly string[] All = new string[]
+                    {
+                        "test-image",
+                    };
                 }
             }
         }
@@ -206,6 +333,18 @@
         public static class TestPack02
         {
             public const string Name = "TestPack02";
+            
+            public static readonly string[] AllNames = new string[]
+            {
+                "TestImage",
+                "TestImage02",
+            };
+            
+            public static readonly string[] AllAddresses = new string[]
+            {
+                "TestImage",
+                "TestImage02",
+            };
             
             public static class TestImage
             {
@@ -216,7 +355,11 @@
                     public const string TestImage = "test-image";
                     public const string TestLabel = "test-label";
                     
-                    public static readonly string[] All = new string[] { "test-image", "test-label" };
+                    public static readonly string[] All = new string[]
+                    {
+                        "test-image",
+                        "test-label",
+                    };
                 }
             }
 
@@ -229,7 +372,11 @@
                     public const string DownloadAll = "download-all";
                     public const string TestImage = "test-image";
                     
-                    public static readonly string[] All = new string[] { "download-all", "test-image" };
+                    public static readonly string[] All = new string[]
+                    {
+                        "download-all",
+                        "test-image",
+                    };
                 }
             }
         }
@@ -237,6 +384,18 @@
         public static class TestScene
         {
             public const string Name = "TestScene";
+            
+            public static readonly string[] AllNames = new string[]
+            {
+                "Sample_Addressables",
+                "Sample_AddressablesAdditive",
+            };
+            
+            public static readonly string[] AllAddresses = new string[]
+            {
+                "Sample_Addressables",
+                "Sample_AddressablesAdditive",
+            };
             
             public static class SampleAddressables
             {
@@ -246,7 +405,10 @@
                 {
                     public const string TestScene = "test-scene";
                     
-                    public static readonly string[] All = new string[] { "test-scene" };
+                    public static readonly string[] All = new string[]
+                    {
+                        "test-scene",
+                    };
                 }
             }
 
@@ -258,7 +420,10 @@
                 {
                     public const string TestScene = "test-scene";
                     
-                    public static readonly string[] All = new string[] { "test-scene" };
+                    public static readonly string[] All = new string[]
+                    {
+                        "test-scene",
+                    };
                 }
             }
         }